Rename XYColumn to Db
Delete OleDbTwoPointLine, use DbTwoPointLine now
