+++ SharpMap/Data/Providers/DbTwoPointLine.cs
