+++ UnitTests/Data/Providers/DbPointProviderTest.cs
@@ -7,7 +7,7 @@ using SharpMap.Data.Providers;
 namespace UnitTests.Data.Providers
 {
     [Category("RequiresWindows")]
-    public class XYColumnTwoPointLineProviderTest : ProviderTest
+    public class XYColumnPointProviderTest : ProviderTest
     {
         private System.Data.Common.DbConnection _connection; // Keep connection active; when closed, the in-memory database is dropped
 
@@ -27,27 +27,18 @@ namespace UnitTests.Data.Providers
             using (var cmd = conn.CreateCommand())
             {
                 cmd.CommandText =
-                    "CREATE TABLE XYColumnTwoPointLineProviderTest(ID integer primary key, Name text, X1 real, Y1 real, X2 real, Y2 real);";
+                    "CREATE TABLE XYColumnPointProviderTest(ID integer primary key, Name text, X real, Y real);";
                 cmd.ExecuteNonQuery();
             }
 
             using (var cmd = conn.CreateCommand())
             {
                 cmd.CommandText =
-                    "INSERT INTO XYColumnTwoPointLineProviderTest(ID, Name, X1, Y1, X2, Y2) VALUES" +
-                    "   (1, 'T1', 100, 100, 170, 100)," +
-                    "   (2, 'T2', 135, 100, 135, 170)," +
-                    "   (3, 'e1', 180, 130, 180, 170)," +
-                    "   (4, 'e2', 180, 130, 210, 130)," +
-                    "   (5, 'e3', 180, 150, 200, 150)," +
-                    "   (6, 'e4', 180, 170, 210, 170)," +
-                    "   (7, 's1', 220, 130, 250, 130)," +
-                    "   (8, 's2', 220, 130, 220, 150)," +
-                    "   (9, 's3', 220, 150, 250, 150)," +
-                    "   (10, 's4', 250, 150, 250, 170)," +
-                    "   (11, 's5', 250, 170, 220, 170)," +
-                    "   (12, 't1', 260, 130, 290, 130)," +
-                    "   (13, 't2', 275, 130, 275, 170)";
+                    "INSERT INTO XYColumnPointProviderTest(ID, Name, X, Y) VALUES" +
+                    "   (1, 'One', 429012.5, 360443.18)," +
+                    "   (2, 'Two', 429001.59, 360446.98)," +
+                    "   (3, 'Three', 429003.31, 360425.45)," +
+                    "   (4, 'Four', 429016.9, 360413.04)";
                 cmd.ExecuteNonQuery();
             }
 
@@ -60,10 +51,10 @@ namespace UnitTests.Data.Providers
             _connection.Close();
         }
 
-        private XYColumnTwoPointLine CreateProvider()
+        private DbPoint CreateProvider()
         {
-            var p = new XYColumnTwoPointLine(System.Data.SQLite.SQLiteFactory.Instance,
-                "FullUri=file::memory:?cache=shared;ToFullPath=false", "XYColumnTwoPointLineProviderTest", "ID", "X1", "Y1", "X2", "Y2");
+            var p = new DbPoint(System.Data.SQLite.SQLiteFactory.Instance,
+                "FullUri=file::memory:?cache=shared;ToFullPath=false", "XYColumnPointProviderTest", "ID", "X", "Y");
 
             return p;
 
@@ -88,8 +79,8 @@ namespace UnitTests.Data.Providers
             {
                 var numFeatures = 0;
                 Assert.DoesNotThrow(() => numFeatures = p.GetFeatureCount());
-                Assert.AreEqual(13, numFeatures);
-                p.DefinitionQuery = "Name='e1'";
+                Assert.AreEqual(4, numFeatures);
+                p.DefinitionQuery = "Name='Two'";
                 Assert.DoesNotThrow(() => numFeatures = p.GetFeatureCount());
                 Assert.AreEqual(1, numFeatures);
             }
@@ -105,7 +96,7 @@ namespace UnitTests.Data.Providers
                 Assert.IsNotNull(feature);
                 Assert.AreEqual(3, Convert.ToInt32(feature[p.ObjectIdColumn]));
                 Assert.AreEqual(feature.Geometry.Centroid.Coordinate,
-                    new Coordinate(180, 150));
+                    new Coordinate(429003.31, 360425.45));
             }
         }
 
@@ -118,7 +109,7 @@ namespace UnitTests.Data.Providers
                 Assert.DoesNotThrow(() => feature = p.GetGeometryByID(3));
                 Assert.IsNotNull(feature);
                 Assert.AreEqual(feature.Centroid.Coordinate,
-                    new Coordinate(180, 150));
+                    new Coordinate(429003.31, 360425.45));
             }
         }
 
@@ -131,7 +122,7 @@ namespace UnitTests.Data.Providers
                 Assert.DoesNotThrow(() => p.ExecuteIntersectionQuery(p.GetExtents(), fds));
                 Assert.AreEqual(1, fds.Tables.Count);
                 var table = fds.Tables[0];
-                Assert.AreEqual(13, table.Rows.Count);
+                Assert.AreEqual(4, table.Rows.Count);
             }
         }
 
@@ -145,7 +136,7 @@ namespace UnitTests.Data.Providers
                 Assert.DoesNotThrow(() => p.ExecuteIntersectionQuery(ext, fds));
                 Assert.AreEqual(1, fds.Tables.Count);
                 var table = fds.Tables[0];
-                Assert.AreEqual(13, table.Rows.Count);
+                Assert.AreEqual(4, table.Rows.Count);
 
                 var oids = p.GetObjectIDsInView(ext);
 
@@ -165,7 +156,7 @@ namespace UnitTests.Data.Providers
                 Assert.DoesNotThrow(() => p.ExecuteIntersectionQuery(ext, fds));
                 Assert.AreEqual(1, fds.Tables.Count);
                 var table = fds.Tables[0];
-                Assert.AreEqual(13, table.Rows.Count);
+                Assert.AreEqual(4, table.Rows.Count);
 
                 var geoms = p.GetGeometriesInView(ext);
 
@@ -182,31 +173,15 @@ namespace UnitTests.Data.Providers
             {
                 var reader = new NetTopologySuite.IO.WKTReader();
                 var poly = reader.Read(
-                    @"POLYGON ((100 100, 170 100, 170 170, 100 170, 100 100))");
+                    @"POLYGON ((428999.76819468878 360451.93329044303, 428998.25517286535 360420.80827007542,
+429023.1119599645 360406.75878171506, 429004.52340613387 360451.71714446822, 
+429004.52340613387 360451.71714446822, 428999.76819468878 360451.93329044303))");
 
                 var fds = new FeatureDataSet();
                 Assert.DoesNotThrow(() => p.ExecuteIntersectionQuery(poly, fds));
                 Assert.AreEqual(1, fds.Tables.Count);
                 var table = fds.Tables[0];
-                Assert.AreEqual(2, table.Rows.Count);
-            }
-        }
-
-        [Test]
-        public void TestExecuteIntersectionQueryAgainstGeometryIntersectionButNoPointsWithin()
-        {
-            using (var p = CreateProvider())
-            {
-                var reader = new NetTopologySuite.IO.WKTReader();
-                var poly = reader.Read(
-                    @"POLYGON ((101 101, 169 101, 169 169, 101 169, 101 101))");
-
-                var fds = new FeatureDataSet();
-                Assert.DoesNotThrow(() => p.ExecuteIntersectionQuery(poly, fds));
-                Assert.AreEqual(1, fds.Tables.Count);
-                var table = fds.Tables[0];
-                Assert.AreEqual(1, table.Rows.Count);
-                Assert.AreEqual("T2", table.Rows[0][1].ToString());
+                Assert.AreEqual(3, table.Rows.Count);
             }
         }
     }
