+++ UnitTests/UnitTests.csproj
@@ -165,8 +165,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Data\Providers\NtsProviderTests.cs" />
-    <Compile Include="Data\Providers\XYColumnPointProviderTest.cs" />
-    <Compile Include="Data\Providers\XYColumnTwoPointLineProviderTest.cs" />
+    <Compile Include="Data\Providers\DbPointProviderTest.cs" />
+    <Compile Include="Data\Providers\DbTwoPointLineProviderTest.cs" />
     <Compile Include="Layers\CoordinateTransformationTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Converters\WKB\WKBTests.cs" />
